# apps/demo-customer-profile-api/openapi.yaml
openapi: 3.0.3
info:
  title: Demo Customer Profile API
  version: 1.0.0
  description: Customer profile and preference lookup API for the Light Fabric workflow demo.
servers:
  - url: http://127.0.0.1:8085
    description: Local development
  - url: http://demo-customer-profile-api:8085
    description: Demo service discovery name
tags:
  - name: Health
  - name: Customers
paths:
  /health:
    get:
      tags:
        - Health
      operationId: getCustomerProfileHealth
      summary: Get service health
      responses:
        "200":
          description: Service health status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
              examples:
                healthy:
                  value:
                    status: UP
                    service: demo-customer-profile-api
  /customers/{customerId}:
    get:
      tags:
        - Customers
      operationId: getCustomerProfile
      summary: Get a customer profile
      parameters:
        - $ref: "#/components/parameters/CustomerId"
      responses:
        "200":
          description: Customer profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerProfile"
              examples:
                premiumCustomer:
                  value:
                    customerId: CUST-1001
                    displayName: Avery Chen
                    segment: premium
                    state: ON
                    accountStatus: active
        "404":
          description: Customer was not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                customerNotFound:
                  value:
                    code: CUSTOMER_NOT_FOUND
                    message: customer CUST-9999 was not found
  /customers/{customerId}/preferences:
    get:
      tags:
        - Customers
      operationId: getCustomerPreferences
      summary: Get customer preferences
      parameters:
        - $ref: "#/components/parameters/CustomerId"
        - name: channel
          in: query
          required: false
          description: Requested channel context. Defaults to portal when omitted.
          schema:
            type: string
            default: portal
          example: portal
      responses:
        "200":
          description: Customer preferences for the requested channel.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreferencesResponse"
              examples:
                portalPreferences:
                  value:
                    customerId: CUST-1001
                    channel: portal
                    consent: true
                    preferredCategories:
                      - travel
                    preferredContactChannel: portal
        "404":
          description: Customer was not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                customerNotFound:
                  value:
                    code: CUSTOMER_NOT_FOUND
                    message: customer CUST-9999 was not found
  /customers/{customerId}/policies:
    get:
      tags:
        - Customers
      operationId: getCustomerPolicies
      summary: Get auto insurance policies for a customer
      parameters:
        - $ref: "#/components/parameters/CustomerId"
      responses:
        "200":
          description: Customer policy context.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyResponse"
              examples:
                activePolicy:
                  value:
                    customerId: CUST-1001
                    policies:
                      - policyId: POL-AUTO-1001
                        status: active
                        effectiveDate: "2025-01-01"
                        expiryDate: "2027-01-01"
                        province: ON
                        coverages:
                          - coverageType: collision
                            deductible: 500
                            limit: 50000
        "404":
          description: Customer was not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /customers/{customerId}/vehicles/{vehicleId}:
    get:
      tags:
        - Customers
      operationId: getCoveredVehicle
      summary: Get a covered vehicle for a customer
      parameters:
        - $ref: "#/components/parameters/CustomerId"
        - $ref: "#/components/parameters/VehicleId"
      responses:
        "200":
          description: Vehicle coverage context.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
              examples:
                coveredVehicle:
                  value:
                    customerId: CUST-1001
                    vehicleId: VEH-1001
                    vin: DEMO-VIN-1001
                    year: 2022
                    make: Toyota
                    model: RAV4
                    covered: true
        "404":
          description: Customer or vehicle was not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /customers/{customerId}/prior-claims:
    get:
      tags:
        - Customers
      operationId: listPriorClaims
      summary: List prior insurance claims for a customer
      parameters:
        - $ref: "#/components/parameters/CustomerId"
      responses:
        "200":
          description: Prior claim history for risk review.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriorClaimsResponse"
              examples:
                priorClaims:
                  value:
                    customerId: CUST-1001
                    priorClaimCount: 1
                    recentClaimCount: 0
                    claims:
                      - claimId: CLM-OLD-1001
                        incidentDate: "2024-08-12"
                        status: closed
                        paidAmount: 1800
        "404":
          description: Customer was not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  parameters:
    CustomerId:
      name: customerId
      in: path
      required: true
      description: Customer identifier.
      schema:
        type: string
      example: CUST-1001
    VehicleId:
      name: vehicleId
      in: path
      required: true
      description: Vehicle identifier.
      schema:
        type: string
      example: VEH-1001
  schemas:
    HealthResponse:
      type: object
      required:
        - status
        - service
      properties:
        status:
          type: string
          example: UP
        service:
          type: string
          example: demo-customer-profile-api
    CustomerProfile:
      type: object
      required:
        - customerId
        - displayName
        - segment
        - state
        - accountStatus
      properties:
        customerId:
          type: string
          example: CUST-1001
        displayName:
          type: string
          example: Avery Chen
        segment:
          type: string
          example: premium
        state:
          type: string
          example: ON
        accountStatus:
          type: string
          example: active
    PreferencesResponse:
      type: object
      required:
        - customerId
        - channel
        - consent
        - preferredCategories
        - preferredContactChannel
      properties:
        customerId:
          type: string
          example: CUST-1001
        channel:
          type: string
          example: portal
        consent:
          type: boolean
          example: true
        preferredCategories:
          type: array
          items:
            type: string
          example:
            - travel
        preferredContactChannel:
          type: string
          example: portal
    PolicyResponse:
      type: object
      required:
        - customerId
        - policies
      properties:
        customerId:
          type: string
          example: CUST-1001
        policies:
          type: array
          items:
            $ref: "#/components/schemas/Policy"
    Policy:
      type: object
      required:
        - policyId
        - status
        - effectiveDate
        - expiryDate
        - province
        - coverages
      properties:
        policyId:
          type: string
          example: POL-AUTO-1001
        status:
          type: string
          example: active
        effectiveDate:
          type: string
          format: date
          example: "2025-01-01"
        expiryDate:
          type: string
          format: date
          example: "2027-01-01"
        province:
          type: string
          example: ON
        coverages:
          type: array
          items:
            $ref: "#/components/schemas/Coverage"
    Coverage:
      type: object
      required:
        - coverageType
        - deductible
        - limit
      properties:
        coverageType:
          type: string
          example: collision
        deductible:
          type: integer
          format: int32
          example: 500
        limit:
          type: integer
          format: int32
          example: 50000
    Vehicle:
      type: object
      required:
        - customerId
        - vehicleId
        - vin
        - year
        - make
        - model
        - covered
      properties:
        customerId:
          type: string
          example: CUST-1001
        vehicleId:
          type: string
          example: VEH-1001
        vin:
          type: string
          example: DEMO-VIN-1001
        year:
          type: integer
          format: int32
          example: 2022
        make:
          type: string
          example: Toyota
        model:
          type: string
          example: RAV4
        covered:
          type: boolean
          example: true
    PriorClaimsResponse:
      type: object
      required:
        - customerId
        - priorClaimCount
        - recentClaimCount
        - claims
      properties:
        customerId:
          type: string
          example: CUST-1001
        priorClaimCount:
          type: integer
          format: int32
          example: 1
        recentClaimCount:
          type: integer
          format: int32
          example: 0
        claims:
          type: array
          items:
            $ref: "#/components/schemas/PriorClaim"
    PriorClaim:
      type: object
      required:
        - claimId
        - incidentDate
        - status
        - paidAmount
      properties:
        claimId:
          type: string
          example: CLM-OLD-1001
        incidentDate:
          type: string
          format: date
          example: "2024-08-12"
        status:
          type: string
          example: closed
        paidAmount:
          type: integer
          format: int32
          example: 1800
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: CUSTOMER_NOT_FOUND
        message:
          type: string
          example: customer CUST-9999 was not found